share bounty import styled
