+++ site/src/components/Bounty/Import/index.js
@@ -1,5 +1,3 @@
-import styled from "styled-components";
-
 import ConnectWallet from "components/ConnectWallet";
 import Button from "@osn/common-ui/es/styled/Button";
 import { accountSelector } from "store/reducers/accountSlice";
@@ -26,66 +24,7 @@ import InputDescription from "components/Common/Import/InputDescription";
 import InputBountyId from "components/Common/Import/InputBountyId";
 import BountyMeta from "components/Common/Import/BountyMeta";
 import { resolveBountyDetailRoute } from "utils/route";
-
-const Wrapper = styled.div`
-  display: flex;
-  align-items: flex-start;
-  @media screen and (max-width: 900px) {
-    flex-direction: column;
-    align-items: stretch;
-  }
-
-  > :first-child {
-    flex-grow: 1;
-
-    > :not(:first-child) {
-      margin-top: 20px;
-    }
-  }
-
-  > :nth-child(2) {
-    flex: 0 0 300px;
-    @media screen and (min-width: 901px) {
-      margin-left: 20px;
-    }
-    @media screen and (max-width: 900px) {
-      margin-top: 20px;
-    }
-  }
-`;
-
-const Box = styled.div`
-  box-shadow: 0px 4px 31px rgba(26, 33, 44, 0.04),
-    0px 0.751293px 3.88168px rgba(26, 33, 44, 0.03);
-  border: 1px solid #f0f3f8;
-  padding: 32px;
-  background-color: white;
-  @media screen and (max-width: 900px) {
-    padding: 16px;
-    margin: 0 -16px;
-  }
-
-  > :not(:first-child) {
-    margin-top: 20px;
-  }
-
-  > :nth-child(5) {
-    margin-top: 8px;
-  }
-`;
-
-const Main = styled(Box)`
-  button {
-    max-width: 86px;
-    float: right;
-  }
-`;
-
-const Side = styled.div`
-  button {
-    box-sizing: border-box;
-  }
-`;
+import { Wrapper, Box, Main, Side } from "./styled";
 
 export default function ImportBounty() {
   const dispatch = useDispatch();
