+++ site/src/components/Bounty/Import/styled.js
@@ -1,0 +1,61 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  display: flex;
+  align-items: flex-start;
+  @media screen and (max-width: 900px) {
+    flex-direction: column;
+    align-items: stretch;
+  }
+
+  > :first-child {
+    flex-grow: 1;
+
+    > :not(:first-child) {
+      margin-top: 20px;
+    }
+  }
+
+  > :nth-child(2) {
+    flex: 0 0 300px;
+    @media screen and (min-width: 901px) {
+      margin-left: 20px;
+    }
+    @media screen and (max-width: 900px) {
+      margin-top: 20px;
+    }
+  }
+`;
+
+export const Box = styled.div`
+  box-shadow: 0px 4px 31px rgba(26, 33, 44, 0.04),
+    0px 0.751293px 3.88168px rgba(26, 33, 44, 0.03);
+  border: 1px solid #f0f3f8;
+  padding: 32px;
+  background-color: white;
+  @media screen and (max-width: 900px) {
+    padding: 16px;
+    margin: 0 -16px;
+  }
+
+  > :not(:first-child) {
+    margin-top: 20px;
+  }
+
+  > :nth-child(5) {
+    margin-top: 8px;
+  }
+`;
+
+export const Main = styled(Box)`
+  button {
+    max-width: 86px;
+    float: right;
+  }
+`;
+
+export const Side = styled.div`
+  button {
+    box-sizing: border-box;
+  }
+`;
